+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/GetMe.kt
@@ -1,6 +1,10 @@
 package com.github.insanusmokrassar.TelegramBotAPI.extensions.api
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
-import com.github.insanusmokrassar.TelegramBotAPI.requests.GetMe
+import com.github.insanusmokrassar.TelegramBotAPI.extensions.api.bot.getMe
 
-suspend fun RequestsExecutor.getMe() = execute(GetMe())
+@Deprecated(
+    "Replaced",
+    ReplaceWith("getMe", "com.github.insanusmokrassar.TelegramBotAPI.extensions.api.bot.GetMeKt.getMe")
+)
+suspend fun RequestsExecutor.getMe() = getMe()